Tests/LibWeb/Text/expected/HTML/canvas-toDataURL-toBlob-origin-clean.txt

untainted toDataURL ok: true
tainted toDataURL: SecurityError
tainted toBlob: SecurityError
tainted getImageData: SecurityError
